Refactoring the server shutdown API.
Splitting the server startup into two main API: The actual initialization and registration, and the main completion queue loop. Introducing dedicated Shutdown API.

This is useful for the internal end-to-end and stress tests. Also, providing a thin
server wrapper for spawning the completion queue in the separate thread. Please note, we may decide to fold this functionality into the server impl class in the future.

In addition, providing a no-op initial client helper unit test to be extended soon using the APIs refactored here.

PiperOrigin-RevId: 373905469