Support for Bazel container images: Part 1 of N.
Adding the scaffolding for pulling down base images, augmenting them with build artifacts and assets, and publishing those images. These rules do not require / use Docker for pulling, building, or pushing images. Also, the images are deterministic and reproducible.

The basic idea is to implement a collection of shell scripts, or a single one, under the release directory for constructing the core images and pushing them to Docker hub. This initial version implements very basic server and client image assembly.

PiperOrigin-RevId: 376056934